+++ ReleaseNotes.md
@@ -1,5 +1,7 @@
 # Новое в этой версии:
 - Реконструкция поверхностей по облаку точек и информации о наблюдениях
+- Теперь построение графа наблюдений происходит на стороне электроника, по этому появилась возможность на ходу фильтровать ребра графа по весу.
 
 # New in this version:
-- Surface reconstruction for point cloud and observations info
+- Surface reconstruction based on point cloud and observations info
+- Now observations graph is building on Elektronik side and user now can filter graph edges by their weight.